Updated openstack/openstack Project: openstack/tripleo-heat-templates  a2aacec531cdafead04a4f8a3b1b5cdad39d8334
Add glance registry service to disable on upgrade

We've broken the upgrade job because anyone upgrading with the
glance registry deployed (and defined in their *Services parameters)
will try to deploy with the old glance-registry.yaml defined in heat.

Instead we define a template which stops and disables the service on
upgrade.

Closes-Bug: #1659079
Change-Id: I03561954d794afae2be06811375d16611fa45973