Updated openstack/openstack Project: openstack-infra/tripleo-ci  0be2b6c1eed7314e522eef984f9ec73c76d03065
Clean up testenv if Jenkins instance goes away

We sometimes see more testenvs in existence than there are running
Jenkins jobs.  The likely cause of this is that a job gets killed
in some unusual way (maybe a new patch set gets pushed), and the
gear client doesn't get a chance to signal back.

Since we can't rely on anything in the instance to handle this
scenario, let's add a check to the testenv-worker that makes sure
the Jenkins instance is still around, and if not signals the gear
client to proceed and delete the testenv.

Note that this replaces the previously added zuul status check,
which proved ineffective because the instance on which it was
running gets deleted before it has a chance to do anything.

Change-Id: I0270ee2ea1498247e0aeb007f4707f9502af8324