Updated openstack/openstack Project: openstack/python-ironicclient  6702a7de1feafe455752148d2b14ec01e4aaf157
Extend OSC "node list" cmd to fetch nodes without instance UUID

The openstackclient command to get a list of nodes is 'openstack baremetal
node list'. It has the optional argument '--associated'. If this is
specified, it will return nodes that are associated with instances
(ie, node.instance_uuid is set to some UUID). However, there is no way to
filter the nodes to get ones that are NOT associated with instances.

Added "--unassociated" option and unit testcase for this.

Change-Id: Iad60e92335bab4eb4a71aedb4e90d34d0186c5cc
Closes-bug: #1619091