Updated openstack/openstack Project: openstack-infra/project-config  81b7e59e113b421b32b1b90cc6d720a0bf2e8f5b
Bump rh1 max-servers to 70

The move to 65 has gone smoothly, and we should have capacity for
even more jobs.  We could probably go straight to 75 (which is the
most we should run with the current te-broker setup), but larger
jumps in the number of testenvs have caused problems in the past so
I'd prefer to stick with 5 at a time.  Note that 5 extra testenvs
will use more than 150 GB of memory, so this increase is not as
small as it looks.

Change-Id: If83b507ba6ca97fc850d3a9caf570c4950949914