Updated openstack/openstack Project: openstack-infra/project-config  c109abaf5e9dea3fffe0d38c6e4eed54c6bb0a28
mark redis jobs in trove CI as non-voting

this is required to break a logjam in the merges required in trove.

In order to fix this problem, we need the gate working again. For that
we need to merge the requirement change blacklisting novaclient 7.0.0
which won't merge because of an issue with redis.

Change-Id: I0aa8939a8bf02841f17d3de4a0c8739961ad852e