Updated openstack/openstack Project: openstack/keystone  a6adf059f7664ba092f962056ed31eeb833a1a90
Do not call `to_dict` outside of a session context

Do not call `to_dict` outside of a session context as if to_dict
interacts with lazy-loaded relationships it can cause errors. For
the most part these are edge-cases and unlikely to happen.

A couple FIXMEs were added to restructure the calls to allow for
`to_dict` to be moved into a method that will act within a session
context.

Change-Id: I769c2cdea1b08a780093d27cdc70bce9f004017b