Updated openstack/openstack Project: openstack/manila  47507ee873dd17039c8989b339fbdaea7796d5c7
Fix devstack plugin to not depend on private network

A private network is only needed in the devstack plugin in the
case of generic driver, dhss=false. Therefore we move the command
that depends on it into the loop so things don't break when the
network doesn't exist if it's not needed.

Change-Id: If8074e6b7ce5d42430666f23484167d6a7262be5