Updated openstack/openstack Project: openstack/tripleo-puppet-elements  688584c435cdfea11d0e37917a3f433205fef084
Add explicit install of net-snmp

Suddenly net-snmp is missing on our mitaka and newton images.  A
dependency must have changed that stopped pulling it in.  Since we
require that package to be installed for the deployment to succeed,
we should install it explicitly.  This will need to be backported
to those stable branches as well.

Change-Id: Ic81a64ae625e7fe208e3cc7f452124724a42d01a
Closes-Bug: 1659384