Updated openstack/openstack Project: openstack/python-openstackclient  dec06aa89a22c1f57e19eb2882f22e5e607bf5ee
Fix 'mapping set' return value

Without this patch, the command 'openstack mapping set <args>' will,
upon success, print the rules for the updated mapping and exit with
return code 1 (failure). This is a problem for scripts and config
management tools that depend on the return code to validate whether the
operation was successful, since even upon success the command returns a
failing error code. Moreover, the behavior of printing the new value is
completely unlike the behavior of any of the 'set' subcommands for
other entities. This patch normalizes the 'mapping set' command by
omitting any return value in the SetMapping take_action() method. This
way the client will only exit with an error code if an exception is
raised, and not upon normal operation.

Change-Id: I610ec3b2fa7561072346d46e49cfc1ae82130e0d