Updated openstack/openstack Project: openstack/python-openstackclient  43f9370915513a3480142419aa8c8d8b0b8c4526
change assert_show_fields to not fail on new fields

whenever a resource adds a field (which is allowed in our API
guidelines), OSC functional tests fail, because we validate
the resource keys to a hardcoded list.

instead, this change proposes that the logic of
assert_show_fields is flipped around, so our hardcoded list acts
as a minimum set of values that must appear in the resource.

as part of this change, some fields were remove from the constants
since they were not actually in the returned data.

also, delete unused code `assert_show_structure`.

Change-Id: I8c0f0e80ea472f9c7f93c5f1f0ae52048e6cd7da