Updated openstack/openstack Project: openstack/nova  3ef7e7f0ddb8498a5ca86eddb0f8982a6f92f556
nova-manage docs: add cells commands prep

This is the first commit in a series to document the nova-manage cells
commands.

This scope of this first commit is limited to cleaning up the section
titles, and removing the hardcoded list of categories as it is prone to
becoming out of date (and is currently not up-to-date).

Change-Id: I59af436fca739de3d4d5033f7b5eb43a8dff287f