Updated openstack/openstack Project: openstack/nova  9a4e7e2b9e3f853bf9a36f1663a0f519648cf089
Handle unicode when dealing with duplicate aggregate errors during migration

The AggregateNameExists message can be translated and have unicode in it
so we can't cast it to a str in python 2.7 else we'll get a
UnicodeEncodeError. That would break the online data migration for
aggregates along with any other online data migrations that run after
that one which impacts upgrades.

This uses six.text_type rather than str for logging the error and adds
a test to recreate the bug and show that it's fixed.

Change-Id: I040db22ecbb9fabe5bbda8a3d9600cc9f76cb170
Closes-Bug: #1653261