Updated openstack/openstack Project: openstack-infra/tripleo-ci  cd4d0d4a5885ade0553eb1539a1c3501f07302cc
Move Multinode environments to THT

Scenarios already live in THT because they contain some data (like
TripleO services or data binding) that are specific to TripleO releases.
Example: NovaPlacement has been introduced in TriplO Ocata but not
support in Newton. That's the reason we moved scenarios templates and
pingtests in THT, so they can be branched.
This patch does the same for multinode and multinode-3nodes. It moves
the roles-data and pingtest to THT.

Change-Id: I689d85918d78f51951e0c142f5cfb51b5c1f78f5