Updated openstack/openstack Project: openstack/cinder  7a25529d44198dd3835ad10c30469a5608f09efe
Switch ManageableSnaphots & ManageableVolumes list to OVO

Currently, the results from ManageableVolumes & ManageableSnapshots list
are being returned as Dict. This needs to be modeled as OVO as this
is being returned via rpc from the driver to the api layer.

We also change all occurences of List Manageable Volumes & snapshots to ovo

Change-Id: Id63e4c35deec6dccc0ae6a82b004618cd214d96e