Updated openstack/openstack Project: openstack/requirements  cff8bd6ab58856e7bbcdd150f6eebe2204be63ea
requires python-kafka>=1.3.1

oslo.messaging kafka driver development is blocked since 6 months
due to the kafka cap.

The Monasca performance issue when the deprecated API is used have been
fixed in 1.3.1. Monasca have switched to pykafka lib to unblock
projects.

So this change remove the cap and raise the minimal requirement.

http://lists.openstack.org/pipermail/openstack-dev/2016-December/108350.html

Change-Id: Ie87dd12ae5d57e0b975a15a3bc1da638ca0a6678
Depends-On: Ibff3fb906c7f6baee54b1199450f49b4bd174d3c
Was-Change-Id: I5aa71d1bd6f741f533e20e4ccf5f75a7b7ef3fac