Updated openstack/openstack Project: openstack/nova  3ffe4dd6861b9751d15c1485e4515f40705460bb
Add sort_key white list for server list/detail

Currently all the DB columns exposed in the sort_key parameter. That means
the supported sort keys will be changed when the DB schema changed. This
patch will strict the supported sort keys in a limited list.

Only the keys which are in the API representation will be supported. For
avoid to break the user hardly, the keys which aren't supported from now
will be ignored when in the request. For other invalid sort keys, the
400 will be returned, which matches the behavior as before.

The sort_key in the JSON-Schema will be enum. But due to there is no
way to describe ignore few of enum values in the schema. So we put all
ignored keys in the enum also. Then we will filter those keys out in the
python code. For readability, we put all the ignored sort keys in the
variable 'SERVER_LIST_IGNORE_SORT_KEY'.

partial implement of bp add-whitelist-for-server-list-filter-sort-parameters

Change-Id: I4af9b0eb2166ce14bde4ef4f8a1f9e9707f2c5cf