Updated openstack/openstack Project: openstack/keystone  169e66ab8800148c4052a46d2cb321af33e44f77
Catch potential SyntaxError in federation mapping

When using the 'groups' keyword in a federation mapping, the value
passed in the assertion map be a simple string with a space. For
example, "ALL USERS". This results in ast.literal_eval() raising
a SyntaxError and not ValueError, which bubbles up to the API as
an uncaught 500 Internal Server Error.

Change-Id: I61f93a6c54b62ba8719d2603f93dc18c33b581ce
Closes-Bug: #1629446