Updated openstack/openstack Project: openstack/cinder  a53068f23c2283f4dca037d43f274e752ffa3e40
Don't drop the volume type's extra spec at DB layer

Now the volume type's extra spec will be always dropped at
DB layer if the request is non admin. But this behavior is
controled by policy. We should return the extra specs to API and
let the policy check to decide whether drop it or not.

Change-Id: I1130a53a02da7aa4b8eb0587186331166d2b9bc1
Closes-bug: #1648717