Updated openstack/openstack Project: openstack-infra/project-config  6976a69b01ebf6cfbd34b52491cc74d406094f63
Fix rally job of zaqar

Recently, we added the Zaqar-UI in Zaqar's devstack when Horizon is
enabled. This is causing the error in Zaqar's Rally job as below:

The /opt/stack/new/zaqar-ui project was not found; if this is a gate
job, add the project to the $PROJECTS variable in the job definition.

This patch is going to fix the issue by add zaqar-ui to the $PROJECTS
variable.

Change-Id: I842501acf1c710bb97eee9ecfbd29758f5a7be26