Updated openstack/openstack Project: openstack/designate  dc34ed43ddbb425fd4b6eb7eeb721daf7d36e147
Fix noop metrics timed decorator

Previously if you did not have the monasca statds available, the @timed
decorator would result in a NoneType error because timed was returning
None instead of acting as a noop function wrapper. This change updates
the NoopTimer to properly handle timed being used as a function.

Change-Id: Ic5543b3503b24aaef960d8b7c0cb38940241cbcb
Closes-Bug: #1659638