Updated openstack/openstack Project: openstack/nova  c1cd098fcb6a918741e1992a5870a198fec6bf85
Fix server group functional test by using all filters

Given we will use the placement API for the next change, we need to verify
all the filters like we already do for other functional tests, and just
use large allocation ratios for cpu and disk instead of disabling the
corresponding filters.

Given we wanted to keep the same test coverage which asserts that instances
are only RAM-bound (because it was only verifying RAMFilter previously),
we are providing a new FakeDriver for having large resources for CPU and disk
so we are sure we are not limited by that.

Change-Id: I92afc017146792968130cf4ad238d67ed1dfa308