Updated openstack/openstack Project: openstack/os-net-config  60f043152dedf04c335a049feedb502b194fbc02
Add check that ovs_extra is passed as list

This patch adds a check that ensures the ovs_extra option, if present,
is passed in as list and raises an InvalidConfigException if not.
It addresses the issue that a user may mistakingly pass the value as
string, which would cause an error later when appending the failure
mode or when formatting the ovs_extra parameter.

Note: Also fixes a sample file in which ovs_extra was passed as string.

Change-Id: I9e8e47390b63d284de10d27b1db2c2cc54c86924
Closes-Bug: #1654196