Updated openstack/openstack Project: openstack-infra/zuul  9a5eada9ad46863d953a3fcc761a8951b334963d
Merger: fully update repo on repo update

When a merger gets a request to update a repo for anything other
than a pre-merge job (eg, ref-updated, periodic, etc), it simply
updates the repository from upstream.  While it did update its
tags and remote branch references, it did not actually update
its local branch heads.  Switch to using the 'reset' method
so that happens (along with the update which was already happening).

This corrects a situation where a periodic or post job might pull
data from a repo on a zuul merger and end up with out of date
information.

Change-Id: Iccc691974ae5efedec9bfa5b553b451a1b5ab686