Updated openstack/openstack Project: openstack/cinder  77ffa914cf5bdb4eb8640a4b4440e42a8e3bb1ee
Switch to using generic groups with Pure driver

This adds implementations of consistent generic volume
groups to the Pure volume drivers. They pretty much
just re-use the CG implementations, and if the group
being created isn't a CG then we bail out and allow
the generic group handling to take care of it.

Change-Id: I02080a1c2eddf93513793ae13adf88d869770a2e
Implements: blueprint pure-generic-volume-groups