Updated openstack/openstack Project: openstack/keystone  f505abb1a837d0d9d1fc6a84ab2ce4616bbe0f7f
Address follow-up comments from previous patchset

There were a couple comments about making minor changes in the patch
set for changing change_password to not require a token. These comments
mentioned fixing the wording on the change_user api-ref note and
adding an additional assertion for one of the added unit tests.

https://review.openstack.org/#/c/404022/

This change corrects the wording in the api-ref note for
change_password about not needing an authentication token and
adds an additional assertion for changing an expired password to
verify that once an expired password is successfully changed, the
user is able to authenticate and create a token.

Change-Id: I8e557d344ee77e0c9c28391d3ef09913bd87fef6