Updated openstack/openstack Project: openstack/glance  7d0d99571a6732e08152657204172fbb0c14fd5f
do not mock private methods of objects from libraries

This patch fixes tests that were using mock to replace a private method
of a library instead of using the public testing interface provided by
that library. The new release of stevedore changes the name of that
private method, which breaks these tests. Because the requirements list
is gated on glance's unit tests, the test failure is preventing other
projects from adopting the new release of stevedore.

Change-Id: Ie9af444e7f8842ffb3e275ede52b802de02e6e99
Signed-off-by: Doug Hellmann <doug@doughellmann.com>