Updated openstack/openstack Project: openstack/horizon  1f1c8c62055144b10aa677d097f7839285b88db3
Support to delete instance from network topology view

When try to delete an instance in network topology view, horizon would
try to load project/instance table in background, then do the delete
row action to the instance.
But the project instance table was paginated, it only load the first
page instances. The delete action would not work if you try to delete
an instance that are not in the first page.
The patch try to fix the bug.

Change-Id: I317bd1ee418d19c075ae3ac8d39563a0514b1795
Closes-Bug: #1597677