Updated openstack/openstack Project: openstack/nova  8557aaeb2f3818de352ce7d0b7bf62d74ddffc4e
Added instance.reboot.error to the legacy notifications

reboot.error is being transformed to use versioned notifications [1].
But in the tests, we are getting an error, as
compute.instance.reboot.error is not added to the allowed legacy
notifications.

However, this legacy notification exists [2] for a long time,
it was updated in 2013. [3]

[1]: https://review.openstack.org/#/c/411791
[2]: https://github.com/openstack/nova/blob/master/nova/compute/manager.py#L3024
[3]: https://github.com/openstack/nova/commit/f35c63df24d8f47927d055729e9ca08c7bc3fe8c

Change-Id: I1ce3deff1894bddb4cde5c3362b1d92816d7b2ca