Updated openstack/openstack Project: openstack-infra/elastic-recheck  a3ee8216ffcf666c75d853ac0125ec47213491f0
Wait until the most recent index is available

When elastic search indexing is behind, and the day has
progressed forward to a new day,  the latest
index is not yet available for use. Exclude it from searches
until it is ready in order to avoid the ElasticHttpNotFoundError.

Add Unit tests for this case as well as for when multiple days
are specified for the search.

Change-Id: Ifd27d1ab21bebcb63b48ea164f425c4a2ac8759c