Updated openstack/openstack Project: openstack-infra/project-config  eec604e827a4df4c6e9b3598d54067836b2c0f0f
barbican: add non-voting tripleo-scenario002

TripleO scenario002 is testing Barbican and boot from encrypted volume.
We think it could be useful to have the feedback on how it works outside
devstack with a non voting job.
We have seen a recent breakage that could have been detected with such a
CI job.

Change-Id: I2bd82f672a6c48a4ed5cd5c128985f101fd9ef2c