Updated openstack/openstack Project: openstack-infra/tripleo-ci  24ccde95cf6371f2d167e7dc8a7b6bb5fd79d7ed
Move readme content to rst format

OpenStack docs are typically written in rst, so this should prevent
people from needing to learn another doc language to contribute to
the tripleo-ci docs.

Part of my motivation is that I'm planning to add some more
content here, and I'd rather not write a bunch of markdown that
would have to be translated to rst if we decide to do more formal
tripleo-ci docs (which we should).

Change-Id: I30f0477b4141b05a72ce1d5bb511488fc1ad867b