Updated openstack/openstack Project: openstack/neutron  cd37b97c294b1cc2f4f6068e90f66c910340f768
Stop passing datapath as 'may_exist' argument

While debugging the related bug, this showed up in the logs:
AddBridgeCommand(datapath_type=None, may_exist=system, name=test-brd8f3648c)

This was caused by the add_bridge command not passing in the datapath
type as named argument so it was taking the positional argument of
may_exist. So the datapath type was being lost whenever this command
was being used.

This fixes it by makeing add_bridge use a named argument for
datapath_type and it also switches it to using the
OVSBridge.create() method to unify the code-paths for bridge
creation.

It's not clear that this was a cause of the bug though, so
this fix is only related.

Change-Id: I5711c29c21741dd847f2c27fb621f723b24cb7fd
Related-Bug: #1627106