Updated openstack/openstack Project: openstack/keystone  d4a1bbda0b29a8011416f83a4a9c5be32669ef2d
Auth Method Handlers now return a response object always

Always return a response object for an AuthMethodHandler. This
allows us to better understand success/failure states from the
Auth Handler Plugins. Instead of relying on arbitrary check of
if user_id is in the AuthContext. This is required for full MFA
support as each plugin will need to report success/failure.

bp: per-user-auth-plugin-reqs
Change-Id: I688b50f61c8a9ce90ca3e7de2471c60f588ef902