Updated openstack/openstack Project: openstack-dev/grenade  99c8c9e998ea9a95845958fd6b1f8dbae9fadef1
Do not wait for nova-compute when ironic is enabled

Ironic should be upgraded before nova according to requirements [0]
using reverse order will lead to nova-compute start failure.
This patch ensures that we do not check nova-compute status in the
nova upgarde phase, when ironic is enabled, as it is expected behaviour.
Ironic will restart nova-compute later.

[0] http://docs.openstack.org/developer/ironic/deploy/upgrade-guide.html#general-upgrades-all-versions

Change-Id: I48dc7649eedb6c6d6753da8ba66fcfccb09acd58
Closes-Bug: #1660581
Related-Bug: #1660646