Updated openstack/openstack Project: openstack-infra/tripleo-ci  77f51c21c090756b6234ef0dd8315511480964b5
tripleo.sh: change the command to retrieve ctlplane subnet cidr

the current command to retrieve the cidr of the ctlplane subnet is
giving inconsistent values. neutronclient somehow changed the behaviour
and is returning a json object that should be parsed.
The change uses osc to retrieve subnet id, and then uses neutronclient
to retrieve the exact cidr value.
This is the most compatible way to do this across all the releases.

Change-Id: I1cb26fbecaf1871401774933642b309c79b35fee